pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
dynamic = ["version"]
description = "Invenio extension for login via edugain."
name = "invenio-edugain"
authors = [
  {name = "Graz University of Technology", email = "info@tugraz.at"},
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["invenio", "edugain"]
license = "MIT AND Apache-2.0"
license-files = [
  "LICENSE",  # MIT license
  "invenio_edugain/assets/semantic-ui/shibboleth-embedded-ds-*/LICENSE*",  # Apache 2.0 license
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "invenio-celery>=2.0.0,<3.0.0",
  "invenio-db>=2.0.0,<3.0.0",
  "invenio-jobs>=4.0.0,<5.0.0",
  "pysaml2>=7.5.0",
  "validators>=0.21",
]

[project.optional-dependencies]
tests = [
  "invenio-app>=2.0.0,<3.0.0",
  "invenio-db[postgresql]>=2.0.0,<3.0.0",
  'lxml>=4.5.2',
  "pytest-black>=0.6.0",
  "pytest-invenio>=3.0.0",
  "ruff>=0.9.6",
]
elasticsearch7 = [
  "invenio-search[elasticsearch7]>=3.0.0,<4.0.0",
]
opensearch2 = [
  "invenio-search[opensearch2]>=3.0.0,<4.0.0",
]

[project.entry-points."invenio_assets.webpack"]
invenio_edugain = "invenio_edugain.webpack:edugain"

[project.entry-points."invenio_base.api_apps"]
invenio_edugain = "invenio_edugain:InvenioEdugain"

[project.entry-points."invenio_base.apps"]
invenio_edugain = "invenio_edugain:InvenioEdugain"

[project.entry-points."invenio_base.blueprints"]
invenio_edugain = "invenio_edugain.views:create_blueprint"

[project.entry-points."invenio_base.finalize_app"]
invenio_edugain = "invenio_edugain.ext:finalize_app"

[project.entry-points."invenio_celery.tasks"]
invenio_edugain = "invenio_edugain.tasks"

[project.entry-points."invenio_db.alembic"]
invenio_edugain = "invenio_edugain:alembic"

[project.entry-points."invenio_db.models"]
invenio_edugain = "invenio_edugain.models"

[project.entry-points."invenio_jobs.jobs"]
ingest_idp_data = "invenio_edugain.jobs:IngestIdPDataJob"

[project.urls]
Repository = "https://github.com/tu-graz-library/invenio-edugain"


[tool.hatch.build]
exclude = [
  "run-tests.sh",
  "tests/**",
]

[tool.hatch.version]
path = "invenio_edugain/__init__.py"


[tool.isort]
profile = "black"


[tool.pytest.ini_options]
addopts = "--black --cov=invenio_edugain --cov-report=term-missing"
testpaths = "tests invenio_edugain"


[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "D203", "D211", "D212", "D213",
  "E501",
  "FA100", "FA102",
  "FIX002",
  "INP001",
  "N999",
  "PERF203",
  "PLR0913",
  "S101",
  "TC002",
  "TD002", "TD003",
  "TID252",
  "UP009",
]
fixable = [
  "RUF022",  # unsorted __all__
]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = [
  "ANN201",
]
